requirements.txt
Flask>=3.1,<4
mysql-connector-python>=9,<10
openpyxl>=3.1,<4
oci>=2,<3
cohere>=5,<6
beautifulsoup4>=4.12,<5
certifi>=2024.2.2
sshtunnel>=0.4,<0.5
paramiko>=3.4,<4.0
